Update main.py
slight comment changes
did we want to completely eliminate testing GPS and checking number of registered users? I find it useless for whoever is using the program. Also, do we need to set user.username and user.userpin within their function? I think we can set these regardless of the function because we set it the same way. Not sure if that makes sense. But we have to set it regardless so do they need to be within their functions.